+++ CHANGES.md
@@ -1,3 +1,8 @@
+**v2.0.1 update**
+
+ - Fixed prev/next arrows not showing up
+ - Added auto hover controls
+
 **v2.0.0 update**
 
  - Update the name of the plugin to reflect the new capabilities